Changes in experiments

+++ experiments/non_stop.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
-methods=(KAHIP METIS)
+methods=(KAHIP)
 partitions=(8 4 2)
-versions=(non_stop)
+versions=(non_stop non_stop_jump)
 workloads=(ycsb_a ycsb_d ycsb_e)
 n_initial_keys=(1000000 1000)
 repartition_intervals=(0)
@@ -9,11 +9,12 @@ reps=1
 
 for w in "${workloads[@]}"; do
 	for initial in "${n_initial_keys[@]}"; do
-		./old workloads/${w}_${initial}.toml
-		mv requests.txt ${w}_${initial}_requests.txt
+		if [ ! -f "${w}_${initial}_requests.txt" ]; then
+			./old workloads/${w}_${initial}.toml
+			mv requests.txt ${w}_${initial}_requests.txt
+		fi
 	done;
 done;
-
 
 for i in $(seq $reps); do
 	echo rep ${i}
@@ -25,9 +26,9 @@ for i in $(seq $reps); do
 						for v in "${versions[@]}"; do
 							mkdir -p output/${w}/${m}
 							echo ${w}_${m}_${initial}_${deltap}_${v}_${p}_${i}
-							if [ ! -f "output/${w}/${m}/${initial}_${deltap}_${v}_${p}_${i}.txt" ]; then
+							if [ ! -f "output/${w}/${m}/${initial}_${deltap}_${v}_${p}_${i}.csv" ]; then
 								./${v} configs/config.toml ${p} ${initial} ${deltap} ${m} ${w}_${initial}_requests.txt > output/${w}/${m}/${initial}_${deltap}_${v}_${p}_${i}.csv
-								mv details.txt output/${w}/${m}/details_${initial}_${deltap}_${v}_${p}_${i}.txt
+								mv details.csv output/${w}/${m}/details_${initial}_${deltap}_${v}_${p}_${i}.csv
 								cp -r output /users/douglasp/
 							fi
 						done;
